+++ lang/en/app.php
@@ -96,6 +96,7 @@ return [
         'fa' => 'Persian',
         'fr' => 'French',
     ],
+    'loading' => 'Loading...',
     'model' => [
         'roles' => 'role',
         'room_types' => 'room type',
@@ -141,7 +142,9 @@ return [
     'user' => 'User',
     'user_name' => 'Name',
     'users' => 'Users',
-    'validation' => [
+    'file' => [
+        'allowed_formats' => 'Allowed file formats: :formats',
+        'max_size' => 'Max. file size: :size',
         'invalid_type' => 'The file type is not allowed.',
         'too_large' => 'The selected file is too large.',
     ],